review: simplify chunked-prefill PR — slim defenses, consolidate module, drop history comments
Style/architecture pass on the chunked-prefill series before review:

- Replace boolean supports_chunked_prefill() with per-walk
  get_chunked_prefill_walks() on submodule_base — mirrors the existing
  can_use_cuda_graphs/get_cuda_graph_configs declaration pattern. Removes
  the hardcoded `if batch.graph_walk != "prefill_text"` special case from
  AREngine._should_chunk_prefill; walk eligibility now lives on the
  submodule that knows it.
- Remove redundant __batched_logits__ defenses in
  AREngine._sample_decode_outputs. The sentinel is provably absent at
  every call site (popped two stack frames up in _execute_batched, never
  used as a top-level key in _execute_sequential).
- Fold mminf/engine/chunked_prefill.py into mminf/engine/ar_engine.py as
  a clearly-labeled section; the orchestrator was 195 lines of pure
  helpers used only by AREngine. Tests updated accordingly.
- Remove perf_testing/chunked_prefill_{smoke,throughput}.py — one-off
  measurement harnesses, not steady-state infra.
- Drop Phase/Tier/Task references from new comments across submodules.py,
  worker.py, micro_scheduler.py — these belong in PR descriptions and
  rot as the codebase evolves.
- Clean stale TODO referencing surfaced max_step_tokens YAML config (now
  in configs/qwen3omni.yaml).
- Delete dead pytest.skip placeholder for walk-level gating (now
  implemented and covered by test_should_chunk_prefill_respects_submodule_walk_declaration).

All 64 chunked-prefill modular tests pass after the refactor.
